+++ src/pytesmo/validation_framework/metric_calculators_adapters.py
@@ -29,17 +29,13 @@ Metric Calculator Adapters change how metrics are calculated by calling
 the `calc_metric` function of the adapted calculator instead of the unadapted
 version.
 """
-import pandas as pd
 
+from pytesmo.time_series.grouping import YearlessDatetime, TsDistributor
 from pytesmo.validation_framework.metric_calculators import (
     PairwiseIntercomparisonMetrics, TripleCollocationMetrics)
 import warnings
 import numpy as np
-from dataclasses import dataclass
-from datetime import datetime
-from typing import Union, List, Tuple, Optional
-from cadati.conv_doy import doy, days_past
-import calendar
+from cadati.conv_doy import days_past
 
 
 def days_in_month(month: int) -> int:
@@ -47,240 +43,6 @@ def days_in_month(month: int) -> int:
     Get number of days in this month (in a LEAP YEAR)
     """
     return days_past[month] - days_past[month - 1]
-
-
-@dataclass
-class GenericDatetime:
-    """
-    Datetime without a year. Only down to second.
-    """
-    month: int
-
-    day: int = 1
-    hour: int = 0
-    minute: int = 0
-    second: int = 0
-
-    @property
-    def __ly(self):
-        return 2400  # arbitrary leap year
-
-    def __ge__(self, other: 'GenericDatetime'):
-        return self.to_datetime(self.__ly) >= other.to_datetime(self.__ly)
-
-    def __le__(self, other: 'GenericDatetime'):
-        return self.to_datetime(self.__ly) <= other.to_datetime(self.__ly)
-
-    def __lt__(self, other: 'GenericDatetime'):
-        return self.to_datetime(self.__ly) < other.to_datetime(self.__ly)
-
-    def __gt__(self, other: 'GenericDatetime'):
-        return self.to_datetime(self.__ly) > other.to_datetime(self.__ly)
-
-    def __repr__(self):
-        return f"****-{self.month:02}-{self.day:02}" \
-               f"T{self.hour:02}:{self.minute:02}:{self.second:02}"
-
-    @property
-    def doy(self) -> int:
-        """
-        Get day of year for this date. Assume leap year!
-        i.e.: 1=Jan.1st, 366=Dec.31st, 60=Feb.29th.
-        """
-        return doy(self.month, self.day, year=None)
-
-    @classmethod
-    def from_datetime(cls, dt: datetime):
-        """
-        Omit year from passed datetime to create generic datetime.
-        """
-        return cls(dt.month, dt.day, dt.hour, dt.minute, dt.second)
-
-    def to_datetime(self, years: Optional[Union[Tuple[int, ...], int]]) \
-            -> Union[datetime, List, None]:
-        """
-        Convert generic datetime to datetime with year.
-        Feb 29th for non-leap-years will return None
-        """
-        dt = []
-
-        for year in np.atleast_1d(years):
-            if not calendar.isleap(year) and self.doy == 60.:
-                continue
-            else:
-                d = datetime(year, self.month, self.day, self.hour,
-                             self.minute, self.second)
-            dt.append(d)
-
-        if len(dt) == 1:
-            return dt[0]
-        elif len(dt) == 0:
-            return None
-        else:
-            return dt
-
-
-class TsDistributor:
-
-    def __init__(self,
-                 dates=None,
-                 date_ranges=None,
-                 generic_dates=None,
-                 generic_date_ranges=None):
-        """
-        Build a data distibutor from individual dates, date ranges, generic
-        dates (without specific year) and generic date ranges.
-
-        Components:
-            - individual datetime objects for distinct dates
-            - generic datetime objects for dates without specific a year
-            - date range / datetime tuple
-                i.e. ALL datetimes between the 2 passed dates (start, end)
-                the start date must be earlier than the end date
-            - generic date range / generic datetime tuple
-                i.e. ALL datetimes between 2 generic dates (for any year)
-
-        Parameters
-        ----------
-        dates : Tuple[datetime, ...] or Tuple[str, ...] or pd.DatetimeIndex
-            Individual dates (that also have a year assigned).
-        date_ranges: Tuple[Tuple[datetime, datetime], ...]
-            A list of date ranges, consisting of a start and end date for each
-            range. The start date must be earlier in time than the end date.
-        generic_dates: Tuple[GenericDateTime,...] or Tuple[datetime...]
-            A list of generic dates (that apply to any year).
-            Can be passed as a list of
-             - GenericDateTime objects
-                    e.g. GenericDateTime(5,31,12,1,10), ie. May 31st 12:01:10
-             - pydatetime objects (years will be ignored, duplicates dropped)
-        generic_date_ranges: [Tuple[GenericDateTime, GenericDateTime], ...]
-            A list of generic date ranges (that apply to any year).
-        """
-
-        self.dates = dates
-        self.date_ranges = date_ranges
-        self.generic_dates = generic_dates
-        self.generic_date_ranges = generic_date_ranges
-
-    def __repr__(self):
-        s = []
-        for var in [
-                'dates', 'date_ranges', 'generic_dates', 'generic_date_ranges'
-        ]:
-            val = getattr(self, var)
-            s.append(f"#{var}={len(val) if val is not None else 0}")
-
-        return f"{self.__class__.__name__}({', '.join(s)})"
-
-    def select(self,
-               df: Union[pd.DataFrame, pd.Series, pd.DatetimeIndex],
-               set_nan=False):
-        """
-        Select rows from data frame or series with mathing date time indices.
-
-        Parameters
-        ----------
-        df: pd.DataFrame or pd.Series
-            Must have a date time index, which is then filtered based on the
-            dates.
-        set_nan: bool, optional (default: False)
-            Instead of dropping rows that are not selected, set their values to
-            nan.
-
-
-        Returns
-        -------
-        df: pd.DataFrame or pd.Series
-            The filterd input data
-
-        """
-        if isinstance(df, pd.DatetimeIndex):
-            idx = df
-        else:
-            idx = df.index
-
-            if not isinstance(idx, pd.DatetimeIndex):
-                raise ValueError(f"Expected a DatetimeIndex, "
-                                 f"but got {type(df.index)}.")
-
-        mask = self.filter(idx)
-
-        if set_nan:
-            df[~mask] = np.nan
-            return df
-        else:
-            return df[mask]
-
-    def filter(self, idx: pd.DatetimeIndex):
-        """
-        Filter datetime index for a TimeSeriesDistributionSet
-
-        Parameters
-        ----------
-        idx: pd.DatetimeIndex
-            Datetime index to split using the set
-
-        Returns
-        -------
-        idx_filtered: pd.DatetimeIndex
-            Filtered Index that contains dates for the set
-        """
-
-        mask = pd.DataFrame(index=idx.copy())
-
-        if self.dates is not None:
-            _idx_dates = idx.intersection(pd.DatetimeIndex(self.dates))
-            mask['dates'] = False
-            mask.loc[_idx_dates, 'dates'] = True
-
-        if self.date_ranges is not None:
-            for i, drange in enumerate(self.date_ranges):
-                start, end = drange[0], drange[1]
-                if start > end:
-                    start, end = end, start
-                mask[f"range{i}"] = (idx >= start) & (idx <= end)
-
-        if self.generic_dates is not None:
-            arrs = np.array([])
-            for d in self.generic_dates:
-                dts = d.to_datetime(np.unique(idx.year))
-                if dts is None:
-                    continue
-                else:
-                    arrs = np.append(arrs, dts)
-            _idx_dates = idx.intersection(pd.DatetimeIndex(arrs))
-            mask['gen_dates'] = False
-            mask.loc[_idx_dates, 'gen_dates'] = True
-
-    # avoid loop like:
-    # cond = ["__index_month == {}".format(m) for m in months]
-    # selection = dat.query(" | ".join(cond)).index
-
-        if self.generic_date_ranges is not None:
-            for i, gdrange in enumerate(self.generic_date_ranges):
-                for y in np.unique(idx.year):
-
-                    if not calendar.isleap(y) and (gdrange[0].doy == 60):
-                        start = GenericDatetime(3, 1)
-                    else:
-                        start = gdrange[0]
-
-                    if (not calendar.isleap(y)) and (gdrange[1].doy == 60):
-                        end = GenericDatetime(2, 28, 23, 59, 59)
-                    else:
-                        end = gdrange[1]
-
-                    start_dt = start.to_datetime(years=y)
-
-                    if end < start:
-                        end_dt = end.to_datetime(years=y + 1)
-                    else:
-                        end_dt = end.to_datetime(years=y)
-
-                    mask[f"gen_range{y}-{i}"] = (idx >= start_dt) & (
-                        idx <= end_dt)
-
-        return mask.any(axis=1, bool_only=True)
 
 
 class SubsetsMetricsAdapter:
@@ -294,7 +56,7 @@ class SubsetsMetricsAdapter:
         TripleCollocationMetrics,
     )
 
-    def __init__(self, calculator, subsets):
+    def __init__(self, calculator, subsets, group_results='tuple'):
         """
         Add functionality to a metric calculator to calculate validation
         metrics for subsets of certain datetimes in a time series
@@ -308,19 +70,28 @@ class SubsetsMetricsAdapter:
         subsets : dict[str, TsDistributor], optional (default: None)
             Define subsets of data. With group names as key and a
             data distributor as values.
+        group_results: str, optional (default: 'tuple')
+            How to group the results.
+            - 'tuple' will group the results by (group, metric)
+            - 'join' will join group and metric name with a '|'
         """
         if not isinstance(calculator, self._supported_metric_calculators):
             warnings.warn(f"Adapting {calculator.__class__} is not supported.")
 
         self.cls = calculator
-
         self.subsets = subsets
+        self.group_results = group_results
+
+        assert group_results in ('tuple', 'join'), \
+            f"Unknown group_results: {group_results}"
 
         # metadata metrics and lon, lat, gpi are excluded from applying
         # seasonally
         self.non_seas_metrics = ["gpi", "lon", "lat"]
-        if self.cls.metadata_template is not None:
-            self.non_seas_metrics += list(self.cls.metadata_template.keys())
+        if hasattr(self.cls, 'metadata_template'):
+            if self.cls.metadata_template is not None:
+                self.non_seas_metrics += list(
+                    self.cls.metadata_template.keys())
 
         all_metrics = calculator.result_template
         subset_metrics = {}
@@ -328,12 +99,21 @@ class SubsetsMetricsAdapter:
         # for each subset create a copy of the metric template
         for name in subsets.keys():
             for k, v in all_metrics.items():
-                if k in self.non_seas_metrics:
-                    subset_metrics[k] = np.array(v)
-                else:
-                    subset_metrics[(name, k)] = np.array(v)
+                subset_metrics[self._genname(name, k)] = np.array(v)
 
         self.result_template = subset_metrics
+
+    def _genname(self, setname: str, metric: str) -> str or tuple:
+        if metric in self.non_seas_metrics:
+            k = f"{metric}"
+        elif self.group_results == 'tuple':
+            k = (f"{setname}", *np.atleast_1d(metric))
+        elif self.group_results == 'join':
+            k = f"{setname}|{metric}"
+        else:
+            raise NotImplementedError(
+                f"Unknown group_results: {self.group_results}")
+        return k
 
     def calc_metrics(self, data, gpi_info):
         """
@@ -354,10 +134,7 @@ class SubsetsMetricsAdapter:
             df = distr.select(data)
             ds = self.cls.calc_metrics(df, gpi_info=gpi_info)
             for metric, res in ds.items():
-                if metric in self.non_seas_metrics:
-                    k = f"{metric}"
-                else:
-                    k = (f"{setname}", *np.atleast_1d(metric))
+                k = self._genname(setname, metric)
                 dataset[k] = res
 
         return dataset
@@ -407,8 +184,8 @@ class MonthsMetricsAdapter(SubsetsMetricsAdapter):
 
         for name, months in sets.items():
             distr = TsDistributor(generic_date_ranges=[(
-                GenericDatetime(m, 1, 0, 0, 0),
-                GenericDatetime(m, days_in_month(m), 23, 59, 59))
+                YearlessDatetime(m, 1, 0, 0, 0),
+                YearlessDatetime(m, days_in_month(m), 23, 59, 59))
                                                        for m in months])
             sets[name] = distr
 
@@ -416,4 +193,4 @@ class MonthsMetricsAdapter(SubsetsMetricsAdapter):
 
 
 if __name__ == '__main__':
-    dt = GenericDatetime(2, 29).to_datetime(2001)
+    dt = YearlessDatetime(2, 29).to_datetime(2001)
